Move JsonEncoder to separate module
